+++ apps/web/src/styles/antd-theme.less
@@ -1,5 +1,6 @@
-@import '~antd/lib/style/themes/default.less'; // <-- you need to import antd variables once in your project
-// @import '~antd/dist/antd.less';
+@import '~antd/lib/style/themes/default.less'; // you need to import antd variables once in your project
+@import '~antd/dist/antd.less'; // Import Ant Design styles
+
 // @import '~highlight.js/styles/monokai-sublime.css';
 // @import '~react-grid-sections/css/styles.css';
 // @import '~react-resizable/css/styles.css';
