+++ libs/frontend/view/templates/Dashboard/DashboardTemplate.tsx
@@ -53,7 +53,6 @@ export const DashboardTemplate = observer(
           >
             <div css={tw`h-full`}>
               <SidebarNavigation />
-              Another one
             </div>
           </Sider>
         )}
