+++ libs/frontend/view/templates/SidebarNavigation/SidebarNavigation.tsx
@@ -1,18 +1,39 @@
+import { PageType } from '@codelab/frontend/abstract/types'
+import { Menu } from 'antd'
+import { useRouter } from 'next/router'
 import React from 'react'
+import { sidebarWidth } from '../Dashboard/constants'
 import { commonSidebarItems } from './commonSidebarItems'
-import { SidebarContainer } from './SidebarNavigationContainer'
 
 export const SidebarNavigation = () => {
+  const router = useRouter()
+
   return (
-    <SidebarContainer
-      fullHeight
-      items={commonSidebarItems}
-      style={{
-        display: 'flex',
-        flexDirection: 'column',
-        justifyContent: 'start',
-        height: '100%',
-      }}
-    />
+    <>
+      <Menu
+        defaultOpenKeys={[]}
+        defaultSelectedKeys={[PageType.AppList]}
+        items={commonSidebarItems}
+        mode="inline"
+        selectedKeys={[router.pathname]}
+        style={{
+          width: `${sidebarWidth}px`,
+          // height: fullHeight ? '100%' : undefined,
+          // maxHeight: '100vh',
+        }}
+      />
+      {/* <Menu */}
+      {/*  defaultOpenKeys={[]} */}
+      {/*  defaultSelectedKeys={[PageType.AppList]} */}
+      {/*  items={items} */}
+      {/*  mode="inline" */}
+      {/*  selectedKeys={[router.pathname]} */}
+      {/*  style={{ */}
+      {/*    width: `${sidebarWidth}px`, */}
+      {/*    // height: fullHeight ? '100%' : undefined, */}
+      {/*    maxHeight: '100vh', */}
+      {/*  }} */}
+      {/*/ > */}
+    </>
   )
 }
