+++ libs/frontend/view/templates/index.ts
@@ -2,4 +2,3 @@ export * from './Dashboard'
 export * from './MainPaneTemplate'
 export * from './SidebarNavigation/commonSidebarItems'
 export * from './SidebarNavigation/SidebarNavigation'
-export * from './SidebarNavigation/SidebarNavigationContainer'
