+++ package.json
@@ -325,7 +325,9 @@
     "madge": "5.0.1",
     "next-compose-plugins": "2.2.1",
     "next-global-css": "1.3.1",
+    "next-plugin-antd-less": "^1.8.0",
     "next-transpile-modules": "^9.0.0",
+    "next-with-less": "^2.0.5",
     "nx": "14.1.7",
     "postcss": "7",
     "prettier": "2.5.1",
